refactor: RecommendationResponse의 생성자 접근 레벨 설정
- 패키지 통일성을 위해 RecommendationAcceptanceTest 를 acceptance 디렉토리로 이동
